% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hilo.R
\name{create_hilo}
\alias{create_hilo}
\title{create_hilo}
\usage{
create_hilo(data = NULL, high = NULL, low = NULL, keepmissing = FALSE)
}
\arguments{
\item{data}{Name of the dataframe that contains the rating scales.  Assumes that
each column is a rating scale of the same scale and will convert them all using
the same cut-off values. Required.}

\item{high}{Value indicating where the High (\"_HI\") cut-off will be.  If the scale
is a 7-point scale where 1 is the lowest and 7 is the highest value, a value
of 6 here would combine values of 6 and 7 assigned a value of 1 to indicate
that the respondent falls into the High (\"_HI\") category and all other values
would be assigned a value of 0 to indicate that the respondent does not fall
into the category. Required.}

\item{low}{Value indicating where the Low (\"_LO\") cut-off will be.  If the scale
is a 7-point scale where 1 is the lowest and 7 is the highest value, a value
of 2 here would combine values of 1 and 2 assigned a value of 1 to indicate
that the respondent falls into the Low (\"_LO\") category and all other values
would be assigned a value of 0 to indicate that the respondent does not fall
into the category. Required.}

\item{keepmissing}{Logical indicating if missing values should be retained.
Default is FALSE, meaning that any missing data would be coded as zero.}
}
\value{
Dataframe containing the names of the scales or variables (column names),
one each for the \"_HI\" dichotomy and one for the \"_LO\" dichotomy.
}
\description{
Creates two dichotomous variables from a single rating scale that
indicates if a respondent's answer is a high value or a low value, based on
cut-offs defined by parameters.  The original variable will be removed and
replaced with two 0,1 variables that are named identically to the original
variable number but given the suffix \"_HI\" to represent the high or positive
end of the rating scale or the suffix \"_LO\" to represent the low or negative
end of the rating scale.
}
\examples{
\dontrun{
if(interactive()){
 qmod::dat1 \%>\%
   dplyr::select(price:smart_tech) \%>\%
   create_hilo(., 4, 2)
 }
}
}
