% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi.R
\name{get_multi}
\alias{get_multi}
\title{get_multi}
\usage{
get_multi(fit = NULL)
}
\arguments{
\item{fit}{The name of the lm or glm fit object from which to pull the values.
Required.}
}
\value{
A tibble object with the multicollinearity statistics in tabular format.
}
\description{
Pull key multicollinearity diagnostics, names VIF and Klein's Rule,
from \code{mctest::imcdiag}.  It reports the variable, the VIF, variable-specific
R-squared value used in Klein's Rule, the overall model R-squared value, the
difference in the R-square values, then flags indicating if if multicollinearity
is detected from the VIF (vif_detect and klein_detect).
}
\examples{
\dontrun{
if(interactive()){
 fit1 <- lm(qmod::mod1, formula=nps~.)
 get_multi(fit1)
 }
}
}
