% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_scores.R
\name{performance_scores}
\alias{performance_scores}
\title{performance_scores}
\usage{
performance_scores(data, high, low)
}
\arguments{
\item{data}{Name of the dataframe that contains the rating scales.  Assumes that
each column is a rating scale of the same scale and will convert them all using
the same cut-off values. Required.}

\item{high}{Value indicating where the Positive cut-off will be.  If the scale
is a 7-point scale where 1 is the lowest and 7 is the highest value, a value
of 6 here would combine values of 6 and 7 into the Positive Category. Required.}

\item{low}{Value indicating where the Negative cut-off will be.  If the scale
is a 7-point scale where 1 is the lowest and 7 is the highest value, a value
of 2 here would combine values of 1 and 2 into the Negative Category.  Any
value that doesn't fall into the Positive or Negative category will be included
in the Neutral category by default. Required.}
}
\value{
Dataframe containing the names of the scales or variables (column names)
and the proportion of the dataset that falls within the three categories defined
by the high and low cutoffs.
}
\description{
Create dataframe that converts rating scales into three categories,
Positive, Neutral, and Negative, appropriate for reporting performance data
during a Key Driver Analysis.
}
\examples{
\dontrun{
if(interactive()){
 qmod::raw1 \%>\%
   dplyr::select(price:smart_tech) \%>\%
   performance_scores(., 4, 2)
 }
}
}
