% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw1}
\alias{raw1}
\title{\code{raw1} is an example key driver dataset but prior to data processing}
\format{
A data frame of 10,408 observations and 13 variables:
\describe{
\item{response_id}{Respondent ID Number}
\item{ltr}{Likelihood to Recommend Rating}
\item{price}{Satisfaction with Price}
\item{billing}{Satisfaction with Billing}
\item{payment_options}{Satisfaction with Payment Options}
\item{customer_service}{Satisfaction with Customer Service}
\item{online_account}{Satisfaction with Online Account}
\item{registration_process}{Satisfaction with Registration Process}
\item{services_products}{Satisfaction with Services Products}
\item{smart_tech}{Satisfaction with Smart Tech}
\item{year}{Year indicator}
\item{fuel_type}{Indicator of Fuel Type}
\item{payment_type}{Indicator of Payment Type}
}
}
\usage{
raw1
}
\description{
Dataset containing all the necessary variables for running a
key driver analysis but prior to any initial data processing.  Once processed,
will result in \code{qmod::dat1}.
}
\details{
LTR is a score ranging from 0 to 10, to be used to construct an NPS score.

All Satisfaction ratings (from price through smart_tech) are based on a 5 point
scale where 1 is Very Unsatisfied and 5 is Very Satisfied.

Remaining variables, year, fuel_type, and payment_type, are nominal variables
used to classify respondents so that missing values can be replaced by the
mean of those groups.
}
\keyword{datasets}
