% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refresh}
\alias{refresh}
\title{Refreshes github install}
\usage{
refresh(pack_list = "qmod", dev = FALSE)
}
\arguments{
\item{pack_list}{A character vector with the names of packages to update. Default: qmod.}

\item{dev}{A logical to indicate if the development version of qmod should be
installed.  Will only work if qmod is the only package listed in the pack_list.
Default: FALSE.}
}
\value{
Nothing, updates package(s).
}
\description{
Convenience function to perform a remotes::install_github for qmod.
}
\examples{
\dontrun{
qmod:::refresh()
qmod:::refresh(dev=TRUE)
}
}
\keyword{internal}
