% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refresh}
\alias{refresh}
\title{Refreshes github install}
\usage{
refresh(dev = FALSE)
}
\arguments{
\item{dev}{A logical to indicate if the development version of qmod should be
installed.  Default: FALSE.}
}
\value{
Nothing, updates package(s).
}
\description{
Convenience function to perform a remotes::install_github for qmod.
}
\examples{
\dontrun{
qmod:::refresh()
qmod:::refresh(dev=TRUE)
}
}
\keyword{internal}
